backend/requirements.txt
Flask==3.1.0
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.1.0
Flask-Login==0.6.3
Flask-Cors==5.0.0
pyodbc==5.2.0
bcrypt==4.2.1
python-dotenv==1.0.1
pytest==8.3.4
